doser_bot/requirements.txt
# Core dependencies
selenium>=4.15.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0

# GUI
customtkinter>=5.2.0
Pillow>=10.0.0

# Web and API
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data generation
faker>=20.0.0
fake-useragent>=1.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorama>=0.4.6
rich>=13.7.0

# CAPTCHA (optional)
# 2captcha-python>=1.1.0

# Proxy support
python-socks>=2.4.0
pysocks>=1.7.1

# Security
cryptography>=41.0.0
